Refactor Part 2
