+++ lsp/src/completion.ts
@@ -1,0 +1,30 @@
+import { CompletionItem, CompletionItemKind, TextDocumentPositionParams } from "vscode-languageserver/node";
+
+export const onCompletion = (_textDocumentPosition: TextDocumentPositionParams): CompletionItem[] => {
+    // The pass parameter contains the position of the text document in
+    // which code complete got requested. For the example we ignore this
+    // info and always provide the same completion items.
+    return [
+        {
+            label: 'TypeScript',
+            kind: CompletionItemKind.Text,
+            data: 1
+        },
+        {
+            label: 'JavaScript',
+            kind: CompletionItemKind.Text,
+            data: 2
+        }
+    ];
+}
+
+export const resolveCompletion = (item: CompletionItem): CompletionItem => {
+    if (item.data === 1) {
+        item.detail = 'TypeScript details';
+        item.documentation = 'TypeScript documentation';
+    } else if (item.data === 2) {
+        item.detail = 'JavaScript details';
+        item.documentation = 'JavaScript documentation';
+    }
+    return item;
+}
